sign is clean
